pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hast"
version = "0.1.0"
description = "AI-native development session manager"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "click>=8.0",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "pytest-bdd>=7.0",
    "litellm>=1.0",
]

[project.optional-dependencies]
pretty = ["rich>=13.0"]
dev = [
    "pytest>=8.0",
    "pytest-xdist>=3.6",
    "pytest-rerunfailures>=14.0",
    "pytest-random-order>=1.1",
    "mutmut>=3.2",
    "ruff>=0.4",
    "mypy>=1.10",
    "pre-commit>=4.0",
]

[project.scripts]
hast = "hast.cli:main"
hast-lab = "hast_lab.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/hast", "src/hast_lab"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.mypy]
python_version = "3.11"
strict = true
